requirements.txt
mysql_connector_repackaged==0.3.1
Pillow==10.0.1
python-telegram-bot==13.4
Requests==2.31.0
sqlitedict==2.1.0
validators==0.22.0
